Add Configuration and Selection modules
